+++ semcore/data-table/__tests__/data-table-cells.browser-test.tsx
@@ -277,4 +277,33 @@ test.describe('Cells', () => {
     }
     await expect(page).toHaveScreenshot();
   });
+
+  test('Verify empty data with selectable rows', async ({ page }) => {
+    const standPath = 'stories/components/data-table/tests/examples/cells-tests/checkbox-in-table-with-no-data.tsx';
+    const htmlContent = await e2eStandToHtml(standPath, 'en');
+    await page.setContent(htmlContent);
+
+    await expect(page).toHaveScreenshot();
+  });
+
+  test('Verify wide indents with selectable rows non compact and compact', async ({ page }) => {
+    const standPath = 'stories/components/data-table/docs/examples/checkbox-in-table.tsx';
+    const htmlContent = await e2eStandToHtml(standPath, 'en', {
+      sideIndents: 'wide',
+    });
+    await page.setContent(htmlContent);
+
+    await test.step('Verify wide for non compact data-tablet', async () => {
+      await expect(page).toHaveScreenshot();
+    });
+
+    await test.step('Verify wide for compact data-tablet', async () => {
+      const htmlContent = await e2eStandToHtml(standPath, 'en', {
+        sideIndents: 'wide',
+        compact: true,
+      });
+      await page.setContent(htmlContent);
+      await expect(page).toHaveScreenshot();
+    });
+  });
 });