+++ semcore/data-table/src/components/Body/Cell.tsx
@@ -7,11 +7,11 @@ import * as React from 'react';
 import type { CellPropsInner, DataTableCellProps } from './Cell.types';
 import { MergedColumnsCell, MergedRowsCell } from './MergedCells';
 import style from './style.shadow.css';
-import { ROW_INDEX } from '../DataTable/DataTable';
+import type { DataTableData } from '../DataTable/DataTable.types';
 
 const DEFAULT_ROW_DURATION = 50;
 
-class CellRoot<UniqKeyType> extends Component<DataTableCellProps<UniqKeyType>, {}, {}, [], CellPropsInner<UniqKeyType>> {
+class CellRoot<Data extends DataTableData, UniqKeyType> extends Component<DataTableCellProps<UniqKeyType>, {}, {}, [], CellPropsInner<Data, UniqKeyType>> {
   static displayName = 'Cell';
   static style = style;
 
@@ -145,6 +145,7 @@ class CellRoot<UniqKeyType> extends Component<DataTableCellProps<UniqKeyType>, {
       isAccordionRow,
       animationExpand,
       style,
+      shadowVertical,
     } = this.asProps;
 
     const cell = row[column.name];
@@ -180,6 +181,7 @@ class CellRoot<UniqKeyType> extends Component<DataTableCellProps<UniqKeyType>, {
         defaultHeight='100%'
         style={style}
         fixed={column.fixed}
+        shadowVertical={column.showShadowVertical ? shadowVertical : undefined}
       >
         <SCell
           ref={this.cellRef}
@@ -202,6 +204,7 @@ class CellRoot<UniqKeyType> extends Component<DataTableCellProps<UniqKeyType>, {
           alignItems={column.alignItems}
           alignContent={column.alignContent}
           justifyContent={column.justifyContent}
+          textAlign={column.textAlign}
         >
           <Children />
         </SCell>