+++ semcore/data-table/src/components/Head/Column.types.ts
@@ -1,5 +1,6 @@
 import type { Property } from 'csstype';
 
+import type { BodyPropsInner } from '../Body/Body.types';
 import type { CellPropsInner } from '../Body/Cell.types';
 import type {
   ColumnGroupConfig,
@@ -43,6 +44,10 @@ export type CommonColumnType = {
    * CSS `justify-content` property
    */
   justifyContent?: Property.JustifyContent;
+  /**
+   * CSS `text-align` property
+   */
+  textAlign?: Property.TextAlign;
 };
 
 export type DTColumn = ColumnItemConfig &
@@ -58,6 +63,8 @@ export type DTColumn = ColumnItemConfig &
     children?: React.ReactNode | React.FC;
 
     gridArea?: string;
+
+    showShadowVertical?: boolean;
   };
 
 export type DataTableColumnProps = CommonColumnType & {
@@ -89,5 +96,7 @@ export type ColumnPropsInner<
   gridTemplateColumns: string[];
   gridTemplateAreas: string[];
   sticky: boolean;
-  onClick: CellPropsInner<UniqKeyType>['onClick'];
+  onClick: CellPropsInner<Data, UniqKeyType>['onClick'];
+  shadowVertical?: BodyPropsInner<Data, UniqKeyType>['shadowVertical'];
+  scrollDirection?: 'horizontal' | 'vertical' | 'both';
 };