+++ semcore/data-table/src/components/Head/Head.types.ts
@@ -1,4 +1,5 @@
-import type { DTColumn } from './Column.types';
+import type { ColumnPropsInner, DTColumn } from './Column.types';
+import type { BodyPropsInner } from '../Body/Body.types';
 import type { CellPropsInner } from '../Body/Cell.types';
 import type { DTRow } from '../Body/Row.types';
 import type { DataTableData, DataTableProps, DTUse } from '../DataTable/DataTable.types';
@@ -25,6 +26,9 @@ export type DataTableHeadProps = {
 
   /** Animation duration for a top property. When it's set adds animation for top changes */
   animationDuration?: number;
+
+  /** Outer ref for the header */
+  ref?: React.Ref<HTMLDivElement>;
 };
 
 export type HeadPropsInner<
@@ -41,7 +45,7 @@ export type HeadPropsInner<
   onSortChange?: DataTableProps<Data, UniqKey, UniqKeyType>['onSortChange'];
   getI18nText: (key: string) => string;
   uid: string;
-  ref: React.RefObject<HTMLDivElement>;
+  ref: React.Ref<HTMLDivElement>;
 
   gridAreaGroupMap: Map<number, string>;
   gridTemplateColumns: string[];
@@ -56,5 +60,8 @@ export type HeadPropsInner<
   getFixedStyle: (
     cell: Pick<DTColumn, 'name' | 'fixed'>,
   ) => [side: 'left' | 'right', style: string | number] | [side: undefined, style: undefined];
-  onCellClick: CellPropsInner<UniqKeyType>['onClick'];
+  onCellClick: CellPropsInner<Data, UniqKeyType>['onClick'];
+  shadowVertical?: BodyPropsInner<Data, UniqKeyType>['shadowVertical'];
+  scrollDirection?: ColumnPropsInner<Data, UniqKey, UniqKeyType>['scrollDirection'];
+
 };