+++ stories/components/d3-chart/tests/radar-chart.stories.tsx
@@ -1,12 +1,19 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
 
-import BasicUsageExample from './examples/radar-chart/basic-usage';
+import BasicUsageExample, { defaultProps as BasicUsageProps } from './examples/radar-chart/basic-usage';
 const meta: Meta = {
   title: 'Components/d3Charts/Tests/Radar-Chart',
 };
 
 export default meta;
 
-export const BasicUsage: StoryObj = {
+export const BasicUsage: StoryObj<typeof BasicUsageProps> = {
   render: BasicUsageExample,
+  argTypes: {
+    showLegend: {
+      control: 'select',
+      options: [true, false, undefined],
+    },
+  },
+  args: BasicUsageProps,
 };