+++ semcore/carousel/CHANGELOG.md
@@ -2,11 +2,17 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.1.7] - 2025-08-29
+
+### Changed
+
+* Type description for `CarouselProps`.
+
 ## [16.1.6] - 2025-08-07
 
 ### Changed
 
-- Type description for `CarouselProps`.
+- Version patch update due to children dependencies update (`@semcore/icon` [16.4.0 ~> 16.5.0], `@semcore/button` [16.0.5 ~> 16.0.6], `@semcore/modal` [16.1.5 ~> 16.1.6], `@semcore/breakpoints` [16.0.5 ~> 16.0.6], `@semcore/base-components` [16.1.2 ~> 16.2.0]).
 
 ## [16.1.5] - 2025-07-23