+++ semcore/d3-chart/CHANGELOG.md
@@ -2,11 +2,17 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [16.1.1] - 2025-08-07
+## [16.1.2] - 2025-08-29
 
 ### Changed
 
 - Type description for `PlotProps`/`ReferenceLineProps`/`AreaChartProps`/`BarChartProps`/`BubbleChartProps`/`CigaretteChartProps`/`DonutChartProps`/`HistogramChartProps`/`LineChartProps`/`RadialTreeProps`/`ScatterPlotChartProps`/`VennChartProps`/`CompactHorizontalBarChartProps`.
+
+## [16.1.1] - 2025-08-08
+
+### Fixed
+
+- Legend hides for single item even when `showLegend=true`.
 
 ## [16.1.0] - 2025-07-23