+++ semcore/fullscreen-modal/CHANGELOG.md
@@ -2,11 +2,17 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [16.1.6] - 2025-08-07
+## [16.1.7] - 2025-08-29
 
 ### Changed
 
 - Type description for `FullscreenModalHeaderProps`.
+
+## [16.1.6] - 2025-08-07
+
+### Changed
+
+- Updated styles to reflect internal `Modal` changes preventing unexpected closing during scrollbar interaction.
 
 ## [16.1.5] - 2025-07-23