+++ semcore/modal/CHANGELOG.md
@@ -2,11 +2,17 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [16.1.6] - 2025-08-07
+## [16.1.7] - 2025-08-29
 
 ### Changed
 
 - Type description for `ModalProps`.
+
+## [16.1.6] - 2025-08-07
+
+### Fixed
+
+- `Modal` hides after scrollbar interaction.
 
 ## [16.1.5] - 2025-07-23
 
@@ -18,7 +24,7 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 ### Fixed
 
-- Modal doesn't close on ESC key press when no element inside is focused.
+- `Modal` doesn't close on ESC key press when no element inside is focused.
 
 ## [16.1.3] - 2025-06-23